ATM Program Version 1.0
This is a program is a simple Automated Teller Machine program. It is
designed
to demonstrate some very simple object oriented programming (OOP) in
Python.

Bank.py is the module that is the bank part of the ATM. It can open
accounts,
close accounts, and switch between accounts.

Account.py is the module that handles individual accounts in a bank.

TODO: Add options to allow a user to switch between banks.

@author: Joshua jjc16_1@yahoo.com